Level 0 & Screens
Level 0 Screen has been finished. In the left side of the screen is Time and Oranges left to collect, in the middle is Level name and the Best Time achieved so far. In the right side are buttons for Muting Music and going back to Levels Screen. Three another screens has been added: Main Screen - with buttons Play, About and Quit. Play button is directed to the Levels screen, About to the About screen and Quit quits the app. On leves screen are 8 levels, each is unlocked after finished previoes level. About screen is unfinished for now.